requirements.txt
python-digitalocean
cutie
appdirs
python-box
ruamel.yaml
black
